Organize main.ts
