+++ src/app/components/Budget/TransactionCategoriesForm/index.tsx
@@ -14,7 +14,10 @@ import SubmitButton from '@app/components/common/Form/SubmitButton';
 import { EntitiesContext } from '@app/context/entitiesContext';
 import { CATEGORY_GROUPED_OPTIONS } from '@appConstants/categories';
 import { EVENT_ERROR, EVENT_NEUTRAL, EVENT_SUCCESS } from '@constants/eventStatus';
-import { DB_EDIT_BUDGET_CATEGORY_ACK, DB_REMOVE_BUDGET_CATEGORY_ACK } from '@constants/events';
+import {
+  DB_EDIT_BUDGET_CATEGORY_ACK,
+  DB_REMOVE_BUDGET_CATEGORY_ACK,
+} from '@constants/repositories';
 import { Budget } from '@database/entities';
 import { StatusBarContext } from '@app/context/statusBarContext';
 import { StatusEnum } from '@app/constants/misc';
