+++ src/app/data/balanceStatements.ipc.ts
@@ -1,6 +1,6 @@
 import { ipcRenderer } from 'electron';
 
-import { DB_GET_BALANCE_STATEMENTS } from '@constants/events';
+import { DB_GET_BALANCE_STATEMENTS } from '@constants/repositories';
 
 export default class BalanceStatementsIpc {
   static getBalanceStatements() {
