+++ src/constants/app.ts
@@ -1,0 +1,3 @@
+export const APP_WINDOW_CONTROL = 'appWindowControl';
+export const APP_INFO = 'appInfo';
+export const APP_HAS_SAFE_STORAGE = 'appHasSafeStorage';
