+++ src/constants/imports.ts
@@ -1,0 +1,7 @@
+export const IMPORT_SOURCE_FILE = 'importSourceFile';
+export const IMPORT_SOURCE_FILE_ACK = 'importSourceFileAck';
+export const ANALYZE_SOURCE_FILE = 'analyzeSourceFile';
+export const ANALYZE_SOURCE_FILE_ACK = 'analyzeSourceFileAck';
+export const LOAD_FROM_CANUTIN_FILE = 'loadFromCanutinFile';
+export const LOAD_FROM_OTHER_CSV = 'loadFromOtherCsv';
+export const LOAD_DATA_ACK = 'loadDataAck';
