+++ src/electron/constants/index.ts
@@ -1,4 +1,4 @@
-export const DID_FINISH_LOADING = 'did-finish-load';
+export const ELECTRON_DID_FINISH_LOADING = 'did-finish-load';
 export const ELECTRON_READY = 'ready';
 export const ELECTRON_WINDOW_CLOSED = 'window-all-closed';
 export const ELECTRON_ACTIVATE = 'activate';
